Do home all work.

+++ src/main/java/com/example/student_management/entity/Student.java
@@ -13,20 +13,15 @@ import lombok.NoArgsConstructor;
 public class Student {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String lastName;
     private String firstName;
     private String fathersName;
     private int age;
     private String birthdate;
     private String passportNumber;
-
-
     @ManyToOne // MANY students To ONE Address
     private Address address;
-
-
-
     @ManyToOne
     private Group group;
 }
